feat: logout

+++ backend/internal/handlers/auth/types.go
@@ -27,7 +27,7 @@ type User struct {
 	ID           string `bson:"_id"`
 	RefreshToken string `bson:"refresh_token"`
 	TokenUsed    bool   `bson:"token_used"`
-	Count        int    `bson:"count"`
+	Count        float64    `bson:"count"`
 }
 type LoginRequest struct {
 	Email    string `json:"email"`
